// src/types/api/service/models/CheckPasswordDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MediaBee Service API
 * MediaBee Service API
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 비밀번호 확인
 * @export
 * @interface CheckPasswordDto
 */
export interface CheckPasswordDto {
    /**
     * 확인할 비밀번호
     * @type {string}
     * @memberof CheckPasswordDto
     */
    password: string;
}

/**
 * Check if a given object implements the CheckPasswordDto interface.
 */
export function instanceOfCheckPasswordDto(value: object): value is CheckPasswordDto {
    if (!('password' in value) || value['password'] === undefined) return false;
    return true;
}

export function CheckPasswordDtoFromJSON(json: any): CheckPasswordDto {
    return CheckPasswordDtoFromJSONTyped(json, false);
}

export function CheckPasswordDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): CheckPasswordDto {
    if (json == null) {
        return json;
    }
    return {
        
        'password': json['password'],
    };
}

export function CheckPasswordDtoToJSON(value?: CheckPasswordDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'password': value['password'],
    };
}